move catalog files
