+++ service/common/src/main/java/org/apache/polaris/service/catalog/iceberg/SupportsCredentialDelegation.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.polaris.service.catalog;
+package org.apache.polaris.service.catalog.iceberg;
 
 import java.util.Map;
 import java.util.Set;
